Remove unused makeArray function
hasDup:
Only called in getMessages() with result of Array#some,
which makes an array.

unique:
Called in getMessages() with result of Array#filter,
which makes an array.
Called in watch() with a value that passed is.array() check.

composeMessage:
Called makeArray() on `names`, which is a private array that
cannot be replaced by user code.