Fix MaxListenersExceededWarning
This follows-up on d56a4cd967, which I accidentally passed
`counter` as the third parameter to watchDirectory(), which is
meant for `fn`.

The consequence is that when watching directories recursively
on Linux, it will call `self.on('change', fn)` multiple times.

In the Travis CI output for Linux, it now says:

> (node:4378) MaxListenersExceededWarning:
> Possible EventEmitter memory leak detected.
> 11 change listeners added.

This is not a failure, but it is a good warning that helped
expose this bug.