Do not close Watcher when there are still subdirs remaining
The callback in Watcher#add calls close() on every kind of
removal, including removals of regular files and  sub directories.

The rest of this function accounts for that correctly, by using
the fullPath to find the associated child watcher, and only
closing that specific one. (And doing nothing, if there is no
matching child watcher, e.g. if it was just a regular file being
removed.)

However, the last lines of the function still unconditionally
closed the main Watcher, which means events would still come
through to the internal debounced callback, but then caught
by the self.isClosed() check and not reported to the user.

Fixes #75.