test: Ignore unrelated events in "should watch directories inside a directory" test
Capture the relevant events and assert afterward. Similar reasons as
079c6123578, but in particular because when testing the Linux-falback
on macOS, this test was actually failing due to a bug in Node.js
(it's bugs all the way down...). The correct event is emitted for
home/c/, but Node.js also emits a phantom event for home/c/c/ before
it, which we need to ignore for the purposes of this test.

Also, move the watcher.close() logic from beforeEach to afterEach.

This way, the watcher of a previous test is closed before the
next test has its tree re-created. This further avoids would-be
test failures where otherwise the old watcher is still getting
events. This is currently fine given that we debounce by default
and absorb the emitted events in the self.isClosed() check, but
it makes debugging a little bit easier if they don't happen in
even internally.